refactor(daml.utils): move ...daml.utils classes to ...daml
