+++ sawtooth-daml-common/src/main/java/com/blockchaintp/sawtooth/daml/Namespace.java
@@ -9,7 +9,7 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 ------------------------------------------------------------------------------*/
-package com.blockchaintp.sawtooth.daml.util;
+package com.blockchaintp.sawtooth.daml;
 
 import java.util.ArrayList;
 import java.util.List;
