+++ sawtooth-daml-common/src/main/java/com/blockchaintp/sawtooth/daml/package-info.java
@@ -13,4 +13,4 @@
 /**
  * Contains common library utility methods and classes for sawtooth-daml.
  */
-package com.blockchaintp.sawtooth.daml.util;
+package com.blockchaintp.sawtooth.daml;
